trigger fetch when worker freed up
Tune the client to be more aggressive about fetching when it either just
fetched a full batch of jobs, or when it skipped its previously
triggered fetch because it was already full. Rather than immediately
calling the fetch limiter as in the previous commit, note this situation
with a boolean so that we can trigger the fetch limiter _after_ a worker
slot has opened up.

This should bring more consistent throughput to poll-only mode and in
cases where there is a backlog of existing jobs but new ones aren't
being actively inserted.

This will result in increased fetch load on many installations, with the
benefit of increased throughput. As before, `FetchCooldown` still limits
how frequently these fetches can occur on each client and can be
increased to reduce the amount of fetch querying.

Finally, don't run the fetch query if there are no worker slots
available.