fix over-aggressive fetching
In ce141fa / #664, the client was tuned in an attempt to make it more
aggressive about fetching more jobs when it has just fetched a full
batch of jobs. Unfortunately, a conditional was missed and it ended up
fetching aggressively any time a single job was fetched.

This adds the correct conditional so that the
`fetchWhenSlotsAreAvailable` flag only gets set when (a) the previous
fetch was full, or (b) when the previous fetch was skipped due to being
already full of jobs.