Update release instructions to remove section on separate CI release (#711)
One nice thing about the use of a Go workspace is that it's no longer
necessary to use `replace` directives in the River CLI module in
`./cmd/river`, but it's still able to use latest code from the other
modules. This means that we no longer need the finnicky two-phase
release process that was previously necessary for the CLI.

Here, correct development instructions to detail only the single unified
release flow. Note the use of `[skip ci]` so as not to pollute the Go
Module cache with the release PR's CI run.