rivertest: implement a Worker type to simplify testing
The `rivertest.Worker` type makes it much simpler to test River workers
with a realistic execution environment without needing to touch the
database at all. The type leverages a `river.Config` to allocate an
inner `Client` and to enable jobs to be easily created with the normal
default values, along with full access to features like middleweare. The
execution context also contains all the normal stuff that a real context
in River might, such as the client and a timeout if configured.