make metadata updates even if job isn't still running
There is a possibility for a metadata update to be staged either later
in the job after a `river.JobCompleteTx` call, or by a middleware after
the job has completed itself. These updates should still be respected.

To do this, make row updates to jobs upon completion
(`JobSetStateIfRunningMany`) even if the jobs aren't still in the
running state, but in that case _only_ merge the metadata fields with no
other updates made.