expose `Config.Test` with `Time` setting
The new `TestConfig` type will contain test-specific client configs. For
now, its only setting is a `Time` field which contains a `TimeGenerator`
type that can be used to set a synthetic clock for testing.

As part of this, the `TimeGenerator` interface was moved from the
semi-internal `rivershared` package to `rivertype` for the sake of API
guarantees. It was also split into two interaces, one of which was kept
in `rivershared` for use by River-internal tests.

Fixes #751.