rivertest: implement a Worker type to simplify testing (#753)
* sliceutil.DefaultIfEmpty -> FirstNonEmpty

Make this util more useful with variadic args, functioning more
similarly to `valutil.FirstNonZero` by returning the first non-empty
value or else `nil`.

* remove unused internal/clientconfig pkg

* testfactory / insert full: allow setting AttemptedBy

* fix panic on non-existent job in JobCompleteTx

* rivertest: implement a Worker type to simplify testing

The `rivertest.Worker` type makes it much simpler to test River workers
with a realistic execution environment without needing to touch the
database at all. The type leverages a `river.Config` to allocate an
inner `Client` and to enable jobs to be easily created with the normal
default values, along with full access to features like middleweare. The
execution context also contains all the normal stuff that a real context
in River might, such as the client and a timeout if configured.

* panic if JobCompleteTx called within a test worker

This can't succeed if the record isn't in the database, so panic if this
situation occurs within a test worker environment.