allow setting created_at on insert for clock overrides
While the client code was attempting to set `created_at` when the clock
was stubbed, the driver code was not actually utilizing this at all.
This commit makes it possible to set the `created_at` (similar to
`scheduled_at`) using a mocked clock. It also removes the ability to
set `finalized_at`, which has no need to be set at insert (other than
for tests utilizing `JobInsertFull`).