rivertest.Worker: Use job executor, always insert jobs, require tx (#766)
* allow setting created_at on insert for clock overrides

While the client code was attempting to set `created_at` when the clock
was stubbed, the driver code was not actually utilizing this at all.
This commit makes it possible to set the `created_at` (similar to
`scheduled_at`) using a mocked clock. It also removes the ability to
set `finalized_at`, which has no need to be set at insert (other than
for tests utilizing `JobInsertFull`).

* Let test workers call `JobCompleteTx` + test work transaction variants

Follows up #753 to make it possible to call `JobCompleteTx` inside test
worker implementations. The test worker tries to update an equivalent
job in the database's state to `running` so `JobCompleteTx` will be able
to find it when called (it only considers jobs in `running` state).

We also augment the API so that it's workable for users who want to use
test transactions in their tests (which is hopefully the vast majority
of users) by adding transaction variants `WorkTx` and `WorkJobTx`.

* always take tx in rivertest.Worker methods

* always insert jobs for testworker

* use real executor for rivertest worker, alter APIs

Leverage the actual job executor for the `rivertest.Worker` type so that
it closely matches real job executions. Reduce the exposed APIs so it
only has `Work` and `WorkJob`, both of which require a transaction (but
the latter works on existing jobs).

Rather than returning only an error, also return a `WorkResult` struct
that contains additional metadata. Populate it with the final updated
job and the event kind to indicate whether the job completed, errored,
snoozed, cancelled, etc.

---------

Co-authored-by: Brandur <brandur@brandur.org>