Fix error message on unsuccessful client subscribe
Fix a bad error message that I just happened to notice while working on
something else. The message says that the subscribe failed because
`Workers` are not configured, but it's actually contingent on `Queues`
not being configured.

You get the error if `subscriptionManager` is nil:

    if c.subscriptionManager == nil {
        panic("created a subscription on a client that will never work jobs (Workers not configured)")
    }

`subscriptionManager is non-nil if `config.willExecuteJobs`:

    if config.willExecuteJobs() {
        if !driver.HasPool() {
                return nil, errMissingDatabasePoolWithQueues
        }

        client.completer = jobcompleter.NewBatchCompleter(archetype, driver.GetExecutor(), client.pilot, nil)
        client.subscriptionManager = newSubscriptionManager(archetype, nil)

And `config.willExecuteJobs` is true if `Queues` is non-empty:

    func (c *Config) willExecuteJobs() bool {
            return len(c.Queues) > 0
    }