propagate testworker errors to caller
If a panic occurs, return a `PanicError` struct containing the trace and
panic cause. Do not use `tb.Fatalf` from within execution except for
unexpected internal issues—instead propagate errors to the user.

Also change the return type to `*WorkResult` since it may be optional
when an internal error occurs.