rivertest.Worker fixes: handle nil config, rename Kind/EventKind, log panic trace (#775)
* rivertest.NewWorker: handle nil config

* rivertest.WorkResult: rename Kind to EventKind

* rivertest.Worker: also log trace when handling panics

* propagate testworker errors to caller

If a panic occurs, return a `PanicError` struct containing the trace and
panic cause. Do not use `tb.Fatalf` from within execution except for
unexpected internal issues—instead propagate errors to the user.

Also change the return type to `*WorkResult` since it may be optional
when an internal error occurs.