allow GOMAXPROCS + 2 test DBs to fix CI flakiness
I'm not entirely sure _why_ this is necessary on GitHub Actions when
it's not at all needed locally. What we continue to see in CI is that
sometimes tests time out on trying to acquire a test DB from the pool of
test DBs, which is currently capped at GOMAXPROCS count. This
_shouldn't_ happen, because that's the same as the number of parallel
tests that should be able to run at once, and so we shouldn't have more
than that number of DBs in demand from concurrent tests.

My hunch is that something is stalling the `TRUNCATE` cleanup actions
and causing one or more tests to be blocked on a DB that's been released
but not yet returned to the pool because it's undergoing cleanup. Simply
increasing the number of DBs to provide a little padding appears to
completely eliminate the staging CI flakiness, _and_ it further speeds
up the `river` package's tests to 20-24 seconds, rather than the 31
seconds we've been seeing since #769 was merged, or the ~40 seconds we
were seeing prior to that.