riverdatabasesql: properly encode bytea unique keys
When the unique key was a binary non-UTF8 []byte, it was generating an
encoding error in the database/sql driver. The fix is to encode it,
similar to what pq does. I couldn't find a way to do this via pq or pgx
without reimplementing this code block, but would love a way to do that
if possible.

Fixes #779.