testdb.Manager: try to recreate pool after cleanup error (#778)
In recent changes from #769 / fb6b3e9, I inadvertently tweaked the logic
of the testdb.Manager to throw away / destroy a resource if it
encountered cleanup errors. Destroying the resource has the side effect
of causing the manager to try to reallocate a new database with a new
number, potentially higher than the number of test databases we've
created.

Instead, try to just make a new pool on the same database, and only
destroy the resource if that fails.