improve driver test coverage for InsertFastMany, fix databasesql unique key issue (#777)
* improve driver test coverage for InsertFastMany unique columns

Make sure we are explicitly testing both intentionally set unique key +
unique states, as well as intentionally omitted.

* failing test for binary unique key on insert

* riverdatabasesql: properly encode bytea unique keys

When the unique key was a binary non-UTF8 []byte, it was generating an
encoding error in the database/sql driver. The fix is to encode it,
similar to what pq does. I couldn't find a way to do this via pq or pgx
without reimplementing this code block, but would love a way to do that
if possible.

Fixes #779.