Job executor: Unmarshal job args within middleware stack if changed
Here, modify the job executor so that instead of exclusively invoking a
work unit's `UnmarshalJSON` implementation before starting to execute
the stack of work middleware, invoke it inside the middleware stack as
well if we detect that `EncodedArgs` on the job row have been modified.
This gives middleware a chance to potentially manipulate encoded args
before they're fully reified to typed job args.

The main driver for this is that I was trying to fully implement an
encrypted args feature, and found it impossible to do so as currently
implemented because any modification a middleware makes to a job row's
encoded args are discarded because typed job args have long since been
unmarshaled before then.