Job executor: Unmarshal job args late, after middlewares have run (#783)
Here, modify the job executor so that instead of exclusively invoking a
work unit's `UnmarshalJSON` implementation before starting to execute
the stack of work middleware, invoke it much later from within the
executor's `doInner` implementation. This gives middlewares the
opportunity to modify args before unmarshaling, thereby enabling use
cases like args encryption.

This does have the downside of having to modify the worker interface's
`Middleware` function from taking a generic job:

    type Worker[T JobArgs] interface {
        Middleware(job *Job[T]) []rivertype.WorkerMiddleware

To taking a `JobRow` instead:

    type Worker[T JobArgs] interface {
        Middleware(job *rivertype.JobRow) []rivertype.WorkerMiddleware

Because the full generic `Job[T]` is not yet available by the time we
need to extract a middleware stack.