Add "hooks" interface in addition to middleware
Here, try an experiment that adds "hooks", a middleware-like concept, but one
which differs in subtle ways, and if done right will unlock functionality that
middleware doesn't make possible, like job-specific actions like encryption.

While experimenting anyway, we also try a variation on middleware configuration
that tries to improve the experience of using hooks to address #788.

The middleware interface is very generic, containing just one sentinel function
so that it's distinct from `interface{}` (and which would make any type
accidentally settable to it):

    type Hook interface {
	// IsHook is a sentinel function to check that a type is implementing Hook
	// on purpose and not by accident (Hook would otherwise be an empty
	// interface). Hooks should embed river.HookDefaults to pick up an
	// implementation for this function automatically.
        IsHook() bool
    }

But from there, we get into specific hook interfaces that run at various phases
in a job's lifecycle. These ones should look familiar from middleware:

    // HookInsertBegin is an interface to a hook that runs before job insertion.
    type HookInsertBegin interface {
        Hook

        InsertBegin(ctx context.Context, params *JobInsertParams) error
    }

    // HookWorkBegin is an interface to a hook that runs after a job has been locked
    // for work and before it's worked.
    type HookWorkBegin interface {
        Hook

        WorkBegin(ctx context.Context, job *JobRow) error
    }

The primary advantage of hooks is that they don't get layered onto the stack,
and especially when inserting rows, don't all need to share a single set of
hooks, letting each job in question run different hooks if desired.

By extension, this also means that unlike middleware, hooks cannot set anything
to context. This is the main practical distinction between the two concepts

By consolidating different phases into a single interface, we let logically
related hooks be configured much more cleanly. Whereas before we needed to
something like this to install a middleware to different lifecycle phases:

    middleware := riverencrypt.NewEncryptMiddleware(riversecretbox.NewSecretboxEncryptor(key))

    config := &river.Config{
        JobInsertMiddleware: []rivertype.JobInsertMiddleware{middleware},
        WorkerMiddleware:    []rivertype.WorkerMiddleware{middleware},
    }

We now get to do this instead:

    config := Config: river.Config{
        Hooks: []rivertype.Hook{
            riverencrypt.NewEncryptHook(riversecretbox.NewSecretboxEncryptor(key)),
        },

This is cleaner code-wise, but it also prevents accidental misuse where say an
encryptor was installed for insertion, but then forgotten for when jobs are
worked.

Similar to `JobArgsWithInsertOpts`, job args can implement `JobArgsWithHooks`
to add hooks for specific job kinds, and which will also tak effect for when
jobs are worked:

    type JobArgsWithHooks interface {
        // Hooks returns specific hooks to run for this job type. These will run
        // after the global hooks configured on the client.
        //
        // Warning: Hooks returned should be based on the job type only and be
        // invariant of the specific contents of a job. Hooks are extract by
        // instantiating a generic instance of the job even when a specific instance
        // is available, so any conditional logic within will be ignored. This is
        // done because although specific job information may be available in some
        // hook contexts like on InsertBegin, it won't be in others like WorkBegin.
        Hooks() []rivertype.Hook
    }